// src/app/modules/bau/taskboard/enums/enums.ts
export enum TaskTrackStaus {
  OnTrack = 1,
  OffTrack = 2,
  NotStarted = 3,
}

export enum TaskStaus {
  Open = 1,
  Closed = 2,
}

export enum TaskImportanceLevel {
  Low = 1,
  Medium = 2,
  High = 3,
}

export enum TaskRecurrentType {
  Daily = 1,
  Weekly = 2,
  Monthly = 3,
  Quarterly = 4,
  BiYearly = 5,
  Yearly = 6,
}